fix: align chat process launch environment
